Updates to everything
Updated entire project to work correctly with a new firebase SDK as well as a new project name
