+++ NavMap/app/build.gradle
@@ -4,11 +4,11 @@ plugins {
 }
 
 android {
-    namespace 'com.example.navmap'
+    namespace 'com.cabandt.navmap'
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.example.navmap"
+        applicationId "com.cabandt.navmap"
         minSdk 25
         targetSdk 32
         versionCode 1
@@ -29,6 +29,7 @@ android {
     }
     buildFeatures {
         viewBinding true
+
     }
 }
 
@@ -41,9 +42,9 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.4.1'
     implementation 'androidx.navigation:navigation-fragment:2.4.1'
     implementation 'androidx.navigation:navigation-ui:2.4.1'
-    implementation 'com.google.firebase:firebase-auth:21.0.7'
-    implementation 'com.google.firebase:firebase-storage:20.1.0'
+    implementation 'com.google.firebase:firebase-auth:21.1.0'
     implementation 'com.google.firebase:firebase-firestore:24.4.0'
+
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
@@ -60,4 +61,5 @@ dependencies {
 
     //places
     implementation 'com.mapbox.mapboxsdk:mapbox-android-plugin-places-v9:0.12.0'
+
 }
