+++ NavMap/app/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools"
+    >
 
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
@@ -17,28 +18,28 @@
         android:theme="@style/Theme.NavMap"
         tools:targetApi="31">
         <activity
-            android:name=".BookmarkViewActivity"
+            android:name="com.cabandt.navmap.BookmarkViewActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".RegisterActivity"
+            android:name="com.cabandt.navmap.RegisterActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".LoginActivity"
+            android:name="com.cabandt.navmap.LoginActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".SplashActivity"
+            android:name="com.cabandt.navmap.SplashActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -51,7 +52,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".NavMain"
+            android:name="com.cabandt.navmap.NavMain"
             android:exported="false"
             android:label="@string/title_activity_nav_main">
             <meta-data
@@ -59,7 +60,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name="com.cabandt.navmap.MainActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
