+++ NavMap/app/src/main/java/com/cabandt/navmap/Bookmark.java
@@ -1,0 +1,47 @@
+package com.cabandt.navmap;
+
+public class Bookmark {
+    String id;
+    String name;
+    String latitude;
+    String longitude;
+
+    public Bookmark(String id, String name, String latitude, String longitude) {
+        this.id = id;
+        this.name = name;
+        this.latitude = latitude;
+        this.longitude = longitude;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getLatitude() {
+        return latitude;
+    }
+
+    public void setLatitude(String latitude) {
+        this.latitude = latitude;
+    }
+
+    public String getLongitude() {
+        return longitude;
+    }
+
+    public void setLongitude(String longitude) {
+        this.longitude = longitude;
+    }
+}
