+++ NavMap/app/src/main/java/com/cabandt/navmap/MainActivity.java
@@ -1,0 +1,32 @@
+package com.cabandt.navmap;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+
+import com.cabandt.navmap.R;
+
+public class MainActivity extends AppCompatActivity {
+
+    Button btnSwitch;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        btnSwitch = findViewById(R.id.switchScreen);
+
+        btnSwitch.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view)
+            {
+                Intent intent = new Intent(MainActivity.this, NavMain.class);
+                startActivity(intent);
+            }
+        });
+    }
+}
