+++ NavMap/app/src/main/res/layout/activity_bookmark_view.xml
@@ -6,7 +6,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BookmarkViewActivity"
+    tools:context="com.cabandt.navmap.BookmarkViewActivity"
     android:background="@color/grey">
 
     <com.mapbox.mapboxsdk.maps.MapView
