+++ NavMap/app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/lightGrey"
-    tools:context=".LoginActivity">
+    tools:context="com.cabandt.navmap.LoginActivity">
 
     <androidx.cardview.widget.CardView
         android:id="@+id/header"
