+++ NavMap/app/src/main/res/layout/activity_register.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/lightGrey"
-    tools:context=".RegisterActivity">
+    tools:context="com.cabandt.navmap.RegisterActivity">
 
     <androidx.cardview.widget.CardView
         android:id="@+id/header"
