+++ NavMap/app/src/main/res/layout/fragment_profile.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/lightGrey"
-    tools:context=".ui.BookmarksFragment"
+    tools:context="com.cabandt.navmap.ui.BookmarksFragment"
     android:paddingBottom="57dp">
 
     <LinearLayout
