+++ NavMap/app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/navigation_explore"
-        android:name="com.example.navmap.ui.ExploreFragment"
+        android:name="com.cabandt.navmap.ui.ExploreFragment"
         android:label="@string/title_explore"
         tools:layout="@layout/fragment_explore" />
 
     <fragment
         android:id="@+id/navigation_bookmarks"
-        android:name="com.example.navmap.ui.BookmarksFragment"
+        android:name="com.cabandt.navmap.ui.BookmarksFragment"
         android:label="@string/title_bookmarks"
         tools:layout="@layout/fragment_bookmarks" />
 
     <fragment
         android:id="@+id/navigation_profile"
-        android:name="com.example.navmap.ui.ProfileFragment"
+        android:name="com.cabandt.navmap.ui.ProfileFragment"
         android:label="@string/title_profile"
         tools:layout="@layout/fragment_profile" />
 </navigation>
